refactor(voice/capture): extract Windows-pair RestartMixin [T1.4 step 8a]
Extract the three Windows-pair restart methods from
``voice/_capture_task.py`` into a new ``RestartMixin`` class at
``voice/capture/_restart_mixin.py`` per CLAUDE.md anti-pattern #16
(god-file split via subpackage with ``__init__.py`` re-exports):

* ``request_exclusive_restart`` (~190 LOC) — APO-bypass reopen in
  WASAPI exclusive mode (v0.20.2 / Bug C downgrade detection).
* ``_reopen_stream_after_device_error`` (~55 LOC) — generic
  reconnect helper.
* ``request_shared_restart`` (~125 LOC) — symmetric revert.

Total: ~370 LOC migrated. ``_capture_task.py`` reduces from 1976
to 1611 LOC.

The mixin declares host-class state attributes (``_running``,
``_tuning``, ``_input_device``, ``_queue``, ``_stream``,
``_sample_rate``, ``_blocksize``, ``_sd_module``, ``_normalizer``,
``_resolved_device_name``, ``_enumerate_fn``, ``_host_api_name``)
and method-via-MRO declarations (``_close_stream``,
``_emit_stream_opened``, ``_signal_consumer_shutdown``,
``_audio_callback``, ``_allocate_ring_buffer``) so mypy strict
accepts every call across the mixin boundary.

Class composition:
  AudioCaptureTask(EpochMixin, RingMixin)
    -> AudioCaptureTask(EpochMixin, RingMixin, RestartMixin)

Step 8b will land the Linux pair
(``request_alsa_hw_direct_restart`` +
``request_session_manager_restart``) on the same mixin in a
follow-up commit, keeping each commit single-responsibility.

Gates: ruff + ruff format + mypy strict (427 files) + bandit
clean. Full pytest 11893 passed; the lone failure
(``test_output_guard::test_none_near_zero``) is a documented
Windows-timer flake (anti-pattern #22) and re-passes in isolation
— not on any path touched by this commit.

Mission: docs-internal/missions/MISSION-voice-final-skype-grade-2026.md §Phase 1.T1.4 step 8a
